third_party/WebKit/LayoutTests/xr/events_deviceconnect.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="../vr/resources/fake-vr-displays.js"></script>
<script src="file:///gen/layout_test_data/mojo/public/js/mojo_bindings.js"></script>
<script src="file:///gen/device/vr/vr_service.mojom.js"></script>
<script src="../vr/resources/mock-vr-service.js"></script>
<canvas id="webgl-canvas"></canvas>
<script>
let fakeDisplays = fakeVRDisplays();

vr_test((t, mockService) => {
  let watcherDone = new Event("watcherdone");
  let eventWatcher = new EventWatcher(t, navigator.xr, ["deviceconnect",
                                                        "watcherdone"]);
  eventWatcher.wait_for(["deviceconnect", "watcherdone"])
    .then( () => {
      t.done();
    });

  // The event should fire when a listener is added even if the devices are not
  // explicity queried with navigator.xr.requestDevice().
  function onDeviceConnect(event) {
    t.step( () => {
      assert_equals(event.device.deviceName, 'Google, Inc. Daydream View');
      navigator.xr.dispatchEvent(watcherDone);
    });
  }
  navigator.xr.addEventListener("deviceconnect", onDeviceConnect, false);

}, [fakeDisplays["Pixel"]],
"Test deviceconnect fires when devices are connected.");

</script>
